uuid_date(): Cast to unsigned long in myerror()
32-bit computers want %u in the format string, 64-bit computers want
%lu. Cast the length to unsigned long to make them agree.

b42d41aa-3fb2-11e9-91e1-4f45262dc9b5

+++ src/uuid.c
@@ -253,7 +253,8 @@ char *uuid_date(char *dest, const char *uuid)
 	strftime(dest, DATE_LENGTH, "%Y-%m-%dT%H:%M:%S", tm);
 	p = dest + strlen(dest);
 	if (p - dest != 19) {
-		myerror("%u: Invalid date length, should be 19", p - dest);
+		myerror("%lu: Invalid date length, should be 19",
+				(unsigned long)(p - dest));
 		return NULL;
 	}
 	snprintf(p, DATE_LENGTH - 19 + 1, ".%07uZ", nano);
